requirements.txt
asgiref==3.5.1
astroid==2.12.12
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.11.0
django-csp==3.7
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.7.1
gunicorn==20.1.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
pipenv==2022.7.24
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
Pygments==2.13.0
PyJWT==2.3.0
pylint==2.15.5
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
requests==2.28.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.2
virtualenv-clone==0.5.7
whitenoise==6.2.0
wrapt==1.14.1
